backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
easyocr
requests
torch
torchvision
torchaudio
pydantic-settings
python-dotenv